Convert all floats to doubles

+++ A6/src/Camera.cpp
@@ -3,7 +3,7 @@
 
 
 
-Camera::Camera(glm::vec4 location, double fovy, int imageWidth, int imageHeight) :
+Camera::Camera(glm::dvec4 location, double fovy, int imageWidth, int imageHeight) :
 	pos(location),
 	fovY(fovy),
 	aspect(imageWidth / imageHeight),
@@ -17,6 +17,6 @@ Ray3D Camera::CreateCameraRay(int rowNum, int colNum) {
 	// u and v are in normalized image coords, -1 to 1
 	double v = (2.0 * ((double)rowNum + 0.5) / (double)imageHeight) - 1.0;
 	double u = (2.0 * ((double)colNum + 0.5) / (double)imageWidth) - 1.0;
-	glm::vec4 rayDir = glm::normalize(glm:: vec4(u * aspect, v, -1.0 * imagePlaneDist, 0.0f));
+	glm::dvec4 rayDir = glm::normalize(glm:: dvec4(u * aspect, v, -1.0 * imagePlaneDist, 0.0f));
 	return Ray3D(pos, rayDir);
 }
