+++ A6/src/Ray3D.h
@@ -3,11 +3,11 @@
 #include <glm/glm.hpp>
 
 struct Ray3D {
-	glm::vec4 start;
-	glm::vec4 dir;
-	Ray3D(glm::vec4 _start, glm::vec4 _dir) : start(_start), dir(_dir) {}
+	glm::dvec4 start;
+	glm::dvec4 dir;
+	Ray3D(glm::dvec4 _start, glm::dvec4 _dir) : start(_start), dir(_dir) {}
 
-	glm::vec4 FindLocAtTime(double time) {
-		return start + ((float)time * dir);
+	glm::dvec4 FindLocAtTime(double time) {
+		return start + (time * dir);
 	}
 };
