+++ A6/src/Sphere.cpp
@@ -5,9 +5,9 @@ using namespace glm;
 
 bool Sphere::IntersectLocal(Ray3D ray, HitResult& outHit, double tMin, double tMax) {
 	// Assume sphere is at origin with radius 1
-	double a = dot(vec3(ray.dir), vec3(ray.dir));
-	double b = 2.0 * dot(vec3(ray.dir), vec3(ray.start));
-	double c = dot(vec3(ray.start), vec3(ray.start)) - 1.0;
+	double a = dot(dvec3(ray.dir), dvec3(ray.dir));
+	double b = 2.0 * dot(dvec3(ray.dir), dvec3(ray.start));
+	double c = dot(dvec3(ray.start), dvec3(ray.start)) - 1.0;
 	// Discriminant^2
 	double d2 = pow(b, 2) - (4 * a * c);
 
@@ -23,7 +23,7 @@ bool Sphere::IntersectLocal(Ray3D ray, HitResult& outHit, double tMin, double tM
 			// Try updating the hit result with whichever t value was the smallest in the range
 			if (outHit.UpdateTMin(timeToUpdate)) {
 				// If it ended up finding a new minT, update the normal value in the hit result
-				vec4 localPos = ray.FindLocAtTime(outHit.t);
+				dvec4 localPos = ray.FindLocAtTime(outHit.t);
 				outHit.nor = normalize(localPos);
 				outHit.nor.w = 0.0; // outHit.nor is a vector, so the w value is always 0
 				return true;
